requirements.txt
Appium-Python-Client==2.11.0
selenium==4.18.1
pytest==7.4.0
pygame==2.5.2
xlrd==2.0.1
gTTS==2.5.1